+++ api-server-nest/src/announcements/dto/announcement-response.ts
@@ -1,6 +1,6 @@
 import {ResponseWrapper} from "../../dto/response-wrapper";
 import isNotNullOrUndefined from "../../utils/is-not-null-or-undefined";
-import cloneAttributes from "../../utils/colone-attributes";
+import cloneAttributes from "../../utils/clone-attributes";
 import {AnnouncementModel} from "../models/announcement.model";
 
 export class AnnouncementResponse extends ResponseWrapper {
