+++ api-server-nest/src/current-events/dto/current-events-response.ts
@@ -1,6 +1,6 @@
 import {ResponseWrapper} from "../../dto/response-wrapper";
 import isNotNullOrUndefined from "../../utils/is-not-null-or-undefined";
-import cloneAttributes from "../../utils/colone-attributes";
+import cloneAttributes from "../../utils/clone-attributes";
 import {CurrentEvent} from "../models/current-event.model";
 
 export class CurrentEventsResponse extends ResponseWrapper {
