+++ api-server-nest/src/venues/dto/venues-response.ts
@@ -1,6 +1,6 @@
 import {ResponseWrapper} from "../../dto/response-wrapper";
 import isNotNullOrUndefined from "../../utils/is-not-null-or-undefined";
-import cloneAttributes from "../../utils/colone-attributes";
+import cloneAttributes from "../../utils/clone-attributes";
 import {VenueModel} from "../models/venue.model";
 
 export class VenuesResponse extends ResponseWrapper {
